requirements.txt
matplotlib==3.8.2
python-dotenv==1.0.1
mysql-connector-python==8.3.0